+++ application/patient-api/src/main/java/com/openclassrooms/mediscreen/patientapi/PatientApiApplication.java
@@ -1,0 +1,13 @@
+package com.openclassrooms.mediscreen.patientapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PatientApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PatientApiApplication.class, args);
+    }
+
+}
