+++ application/patient-api/src/main/java/com/openclassrooms/mediscreen/patientapi/constant/ErrorMessage.java
@@ -1,0 +1,10 @@
+package com.openclassrooms.mediscreen.patientapi.constant;
+
+public final class ErrorMessage {
+
+    public static final String FIELD_IS_MANDATORY = "Field is mandatory";
+
+    public static final String TOO_MUCH_CHARACTERS = "Entry has too much characters";
+
+
+}
