+++ application/patient-api/src/main/java/com/openclassrooms/mediscreen/patientapi/constant/Number.java
@@ -1,0 +1,10 @@
+package com.openclassrooms.mediscreen.patientapi.constant;
+
+public final class Number {
+
+    public static final int FIFTEEN = 15;
+
+    public static final int TWENTY = 20;
+
+    public static final int HUNDRED = 100;
+}
