+++ application/patient-api/src/main/java/com/openclassrooms/mediscreen/patientapi/repository/PatientRepository.java
@@ -1,0 +1,8 @@
+package com.openclassrooms.mediscreen.patientapi.repository;
+
+import com.openclassrooms.mediscreen.patientapi.model.Patient;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PatientRepository extends JpaRepository<Patient, Integer> {
+
+}
