+++ application/patient-api/src/main/resources/application.properties
@@ -1,0 +1,23 @@
+
+server.port=8081
+
+##################### SL4J ##########################################
+
+logging.level.org.springframework=INFO
+logging.file.name= patient-api.log
+logging.file.path=/Mediscreen
+
+
+################### DataSource Configuration ##########################
+
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:3306/demo?serverTimezone=UTC
+spring.datasource.username=root
+spring.datasource.password=password
+
+################### Hibernate Configuration ##########################
+
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.open-in-view=true
+spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
