+++ application/patient-api/src/test/java/com/openclassrooms/mediscreen/patientapi/PatientApiApplicationTest.java
@@ -1,0 +1,13 @@
+package com.openclassrooms.mediscreen.patientapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class PatientApiApplicationTest {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
