+++ application/web-app/src/main/java/com/openclassrooms/mediscreen/webapp/MediscreenApplication.java
@@ -1,0 +1,13 @@
+package com.openclassrooms.mediscreen.webapp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MediscreenApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MediscreenApplication.class, args);
+    }
+
+}
