+++ application/web-app/src/main/java/com/openclassrooms/mediscreen/webapp/constant/Number.java
@@ -1,0 +1,10 @@
+package com.openclassrooms.mediscreen.webapp.constant;
+
+public final class Number {
+
+    public static final int FIFTEEN = 15;
+
+    public static final int TWENTY = 20;
+
+    public static final int HUNDRED = 100;
+}
