+++ application/web-app/src/main/resources/application.properties
@@ -1,0 +1,15 @@
+
+server.port=8080
+
+##################### SL4J ##########################################
+
+logging.level.org.springframework=INFO
+logging.file.name= web-app.log
+logging.file.path=/Mediscreen
+
+################### Thymeleaf #####################
+
+spring.thymeleaf.cache=false
+spring.thymeleaf.enabled=true
+spring.thymeleaf.prefix=classpath:/templates/
+spring.thymeleaf.suffix=.html
