+++ application/web-app/src/main/resources/scripts/schema-demo.sql
