+++ application/web-app/src/main/resources/templates/patients/list.html
