+++ application/web-app/src/test/java/com/openclassrooms/mediscreen/webapp/MediscreenApplicationTest.java
@@ -1,0 +1,13 @@
+package com.openclassrooms.mediscreen.webapp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MediscreenApplicationTest {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
