+++ buildSrc/src/main/kotlin/DependenciesVersion.kt
@@ -1,0 +1,4 @@
+object DependenciesVersions {
+    const val springDependencyManagement = "1.0.11.RELEASE"
+    const val springBoot = "2.5.3"
+}
