+++ core-librairies/models/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    java
+    jacoco
+}
+
+version = "0.0.1-SNAPSHOT"
+
+dependencies {
+    implementation("com.fasterxml.jackson.core:jackson-databind:2.12.4")
+    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.12.4")
+
+}
