--- src/main/java/com/openclassrooms/mediscreen/MediscreenApplication.java
@@ -1,13 +1,0 @@
-package com.openclassrooms.mediscreen;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MediscreenApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MediscreenApplication.class, args);
-    }
-
-}
