--- src/main/java/com/openclassrooms/mediscreen/constant/ErrorMessage.java
@@ -1,10 +1,0 @@
-package com.openclassrooms.mediscreen.constant;
-
-public final class ErrorMessage {
-
-    public static final String FIELD_IS_MANDATORY = "Field is mandatory";
-
-    public static final String TOO_MUCH_CHARACTERS = "Entry has too much characters";
-
-
-}
