--- src/main/java/com/openclassrooms/mediscreen/constant/Number.java
@@ -1,10 +1,0 @@
-package com.openclassrooms.mediscreen.constant;
-
-public final class Number {
-
-    public static final int FIFTEEN = 15;
-
-    public static final int TWENTY = 20;
-
-    public static final int HUNDRED = 100;
-}
