--- src/main/java/com/openclassrooms/mediscreen/controller/HomeController.java
@@ -1,17 +1,0 @@
-package com.openclassrooms.mediscreen.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-
-    public HomeController() {
-
-    }
-
-    @GetMapping("home")
-    public String homePage() {
-        return "/home";
-    }
-}
