--- src/main/java/com/openclassrooms/mediscreen/exception/ElementNotFoundException.java
@@ -1,16 +1,0 @@
-package com.openclassrooms.mediscreen.exception;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class ElementNotFoundException extends RuntimeException {
-
-    /**
-     * Exception to throw when try to find
-     * a not existing element as user account, transfer...
-     * @param message .
-     */
-    public ElementNotFoundException(final String message) {
-        super(message);
-    }
-}
