--- src/main/java/com/openclassrooms/mediscreen/repository/PatientRepository.java
@@ -1,8 +1,0 @@
-package com.openclassrooms.mediscreen.repository;
-
-import com.openclassrooms.mediscreen.model.Patient;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PatientRepository extends JpaRepository<Patient, Integer> {
-
-}
