--- src/main/resources/application.properties
@@ -1,30 +1,0 @@
-
-server.port=8080
-
-##################### SL4J ##########################################
-
-logging.level.org.springframework=INFO
-logging.file.name= sl4j.log
-logging.file.path=/Mediscreen
-
-
-################### DataSource Configuration ##########################
-
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/demo?serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=password
-
-################### Hibernate Configuration ##########################
-
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.open-in-view=true
-spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
-
-################### Thymeleaf #####################
-
-spring.thymeleaf.cache=false
-spring.thymeleaf.enabled=true
-spring.thymeleaf.prefix=classpath:/templates/
-spring.thymeleaf.suffix=.html
