--- src/test/java/com/openclassrooms/mediscreen/MediscreenApplicationTests.java
@@ -1,13 +1,0 @@
-package com.openclassrooms.mediscreen;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MediscreenApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
